Centralized Input System in Game Manager and added improved CardStorage and visuals.

+++ Assets/Scripts/Drawing/Cards/Info/CardInfoSO.cs
@@ -8,5 +8,5 @@ public abstract class CardInfoSO : ScriptableObject {
     [SerializeField] public byte maxUses;
     [SerializeField] public GameObject cardUIPrefab;
 
-    public abstract Card GetCard(byte position);
+    public abstract Card GetCard(int position);
 }
